git-svn-id: file:///home/yoann/SVN/trunk/libprelude@3605 09c5ec92-17d4-0310-903a-819935f44dba

+++ src/include/idmef-value.h
@@ -95,7 +95,7 @@ idmef_value_t *idmef_value_clone(idmef_value_t *val);
 
 idmef_value_t *idmef_value_ref(idmef_value_t *val);
 
-int idmef_value_to_string(idmef_value_t *val, char *buf, int len);
+int idmef_value_to_string(idmef_value_t *val, char *buf, size_t len);
 
 int idmef_value_get(void *res, idmef_value_t *val);
